Moved appXMLInfo to a Models location
